+++ config/EURUSD_M30.h
@@ -5,9 +5,7 @@
 
 // Defines indicator's parameter values for the given pair symbol and timeframe.
 struct Indi_Fractals_Params_M30 : Indi_Fractals_Params {
-  Indi_Fractals_Params_M30() : Indi_Fractals_Params(indi_fractals_defaults, PERIOD_M30) {
-    shift = 0;
-  }
+  Indi_Fractals_Params_M30() : Indi_Fractals_Params(indi_fractals_defaults, PERIOD_M30) { shift = 0; }
 } indi_fractals_m30;
 
 // Defines strategy's parameter values for the given pair symbol and timeframe.
@@ -22,7 +20,7 @@ struct Stg_Fractals_Params_M30 : StgParams {
     signal_close_method = 0;
     signal_close_level = (float)0;
     price_stop_method = 0;
-    price_stop_level = 2;
+    price_stop_level = (float)2;
     tick_filter_method = 1;
     max_spread = 0;
   }
